Add belongs_to :user to Product model, has_many :products to User model, and refactor seeds.rb to require separate seeder files" "
