+++ app/models/product.rb
@@ -1,2 +1,3 @@
 class Product < ApplicationRecord
+    belongs_to :user
 end
