+++ app/models/user.rb
@@ -1,3 +1,4 @@
 class User < ApplicationRecord
     has_secure_password
+    has_many :products
 end
